Actually make the network call
And add more error handling situations

+++ app/src/main/java/com/iulian/iancu/memeweather/data/WeatherResult.kt
@@ -87,7 +87,7 @@ data class Daily(
     val rain: Double,
     @Expose
     @SerializedName("uvi")
-    var uvi: Int
+    var uvi: Double
 )
 
 data class Weather(
